added a text
