+++ code/README.md
@@ -1,19 +1,6 @@
-# Technigo React Starter App
+Happy thoughts app!
 
-This app has been generated using `create-react-app`, then cleaned up a little so we have a project with just the elements we need to get started.
+It was a fun but complicated project with small bits that could easily break the code. React get more and more logic and looking forward to learn and practice more.  
 
-Add components in the `src` folder and use them in `src/app.js` to get started.
+Netlify: 
 
-## Usage
-
-1. On GitHub, click the 'use this template' button to create a new repo on your account
-1. Clone the new repo to your computer
-1. Install required dependencies with npm: `npm install`
-1. Start the project: `npm start`
-
-Alternatively, if you don't want a new git repo on your account;
-
-1. Download & extract the [latest release](https://github.com/Technigo/react-starter/releases/latest) from GitHub
-1. Open the directory in the terminal: `cd /path/to/react-starter`
-1. Install required dependencies with npm: `npm install`
-1. Start the project: `npm start`
