--- code/index.css
@@ -1,103 +1,0 @@
-body {
-  margin: 20px;
-  font-family: "Roboto Mono", monospace;
-  text-align: center;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  box-sizing: border-box;
-  background-color: lightslategray;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
-}
-
-h1 {
-  font-size: 14px;
-  font-size: 15px;
-}
-
-form {
-  display: flex;
-  background-color: #f2f0f0;
-  border: 1px solid;
-  flex-direction: column;
-  width: 400px;
-  padding: 10px;
-  box-shadow: 5px 5px #000;
-  text-align: center;
-  margin: 20px auto 10px;
-
-}
-
-.hearts {
-  padding: 0 15px;
-}
-
-.text-input {
-  margin-bottom: 16px;
-  min-height: 70px;
-}
-.thought-container {
-  display: flex;
-  background-color: #ffff;
-  border: 1px solid;
-  width: 400px;
-  text-align: center;
-  margin: 30px auto;
-  padding: 10px;
-  box-shadow: 5px 5px #000;
-  word-wrap: break-word;
-  flex-wrap: wrap;
-  vertical-align: -webkit-baseline-middle;
-  position: relative;
-  justify-content: space-between;
-}
-
-.thought-message {
-  width: 100%;
-
-}
-
-.submit-button {
-  cursor: pointer;
-  border-radius: 20px;
-  width: 275px;
-  height: 40px;
-  background-color: pink;
-  text-align: center;
-  margin: 0 auto;
-}
-
-.heart-btn {
-  border: none;
-  background: pink;
-  border-radius: 100%;
-  width: 40px;
-  height: 40px;
-  margin-top: 4px;
-}
-
-.heart-emoji {
-  margin-left: -3px;
-}
-
-.like-container {
-  display: flex;
-}
-
-.likes {
-  padding: 0 10px;
-}
-
-.date {
-  width: 50%;
-  text-align: right;
-  margin-top: 16px;
-  justify-items: flex-end;
-  }
-
-
-  
-
